requirements.txt
--find-links https://storage.googleapis.com/jax-releases/jax_releases.html

jaxlib==0.4.14
jax==0.4.14
tensorflow-federated==0.87.0
tensorflow==2.14.1
tensorflow_privacy==0.9.0
paho-mqtt==2.1.0
pandas==2.2.3